Simplify i18n by using a CSS stylesheet
Move styles that were formerly partly in HTML and partly in JS to a
separate CSS file. This simplifies i18n, since it’s enough to set the
directionality at one place, everything else can reference that. It
hopefully also makes it easier to make this extension dark mode
compatible in the future.